move swipable components into recommendation scene. create deck navigator.
