+++ components/recommendations-scene/deck-navigator/styles.js
@@ -1,0 +1,16 @@
+'use strict';
+
+import React, {StyleSheet} from 'react-native';
+
+export default StyleSheet.create({
+  deck: {
+    flex: 1
+  },
+
+  label: {
+    padding: 10,
+    color: '#fff',
+    width: 150,
+    textAlign: 'center'
+  },
+});
